ai-service/requirements.txt
Flask==3.1.3
Flask-Limiter==4.1.1
flask-cors==6.0.2
python-dotenv==1.2.2
requests==2.32.5
redis==5.2.1
sentence-transformers
chromadb
